v2.0.6-pre8 	+ added byte_number() methods to TokenReaderNative 	+ fixed double line-counting bug in mootTokenLexer.ll in presence of tags 	+ added byte-counting code to mootTokenLexer.ll (the template really ought to do this for us)
git-svn-id: file:///home/svn/dev/moot/moot-2.0.6-pre8@2097 4a515de8-beb3-11dc-84d6-2b654028afac

+++ src/libmoot/mootTokenIO.cc
@@ -322,7 +322,7 @@ void TokenReader::carp(const char *fmt, ...)
 	  ((tr_istream
 	    && tr_istream->valid()
 	    && !tr_istream->name.empty()) ? tr_istream->name.c_str() : "(unknown)"),
-	  byte_number(),
+	  static_cast<ssize_t>(byte_number()),
 	  line_number(),
 	  column_number());
 }
